@@ -16,27 +16,29 @@
 		}
 	],
 	"require": {
-		"php": "^7.3|^8",
-		"doctrine/doctrine-bundle": "^1.6.12|^2",
-		"doctrine/orm": "^2.5.2",
-		"symfony/config": "~4.4|~5.3|^6",
-		"symfony/dependency-injection": "~4.4|~5.3|^6",
-		"symfony/framework-bundle": "~4.4|~5.3|^6"
+		"php": "^8",
+		"doctrine/doctrine-bundle": "^2.7",
+		"doctrine/orm": "^2.14",
+		"symfony/config": "~4.4|~5.4|^6",
+		"symfony/dependency-injection": "~4.4|~5.4|^6",
+		"symfony/framework-bundle": "~4.4|~5.4|^6",
+		"symfony/validator": "~4.4|~5.4|^6"
 	},
 	"require-dev": {
-		"doctrine/annotations": "^1.6",
-		"doctrine/data-fixtures": "^1.0.1",
-		"doctrine/dbal": "^2.7|^3",
-		"doctrine/doctrine-fixtures-bundle": "^2.4.1|^3.1",
-		"phpstan/extension-installer": "^1.1",
-		"phpstan/phpstan": "^1.4",
-		"phpstan/phpstan-deprecation-rules": "^1.0",
-		"phpstan/phpstan-strict-rules": "^1.1",
-		"phpstan/phpstan-symfony": "^1.1",
+		"doctrine/annotations": "^1.14",
+		"doctrine/data-fixtures": "^1.6",
+		"doctrine/dbal": "^2.13|^3.5",
+		"doctrine/doctrine-fixtures-bundle": "^3.4",
+		"phpstan/extension-installer": "^1.2",
+		"phpstan/phpstan": "^1.9",
+		"phpstan/phpstan-deprecation-rules": "^1.1",
+		"phpstan/phpstan-strict-rules": "^1.4",
+		"phpstan/phpstan-symfony": "^1.2",
 		"phpunit/phpunit": "^9.5",
-		"symfony/browser-kit": "~4.4|~5.3|^6",
+		"symfony/browser-kit": "~4.4|~5.4|^6",
 		"symfony/phpunit-bridge": "^6",
-		"symfony/yaml": "~4.4|~5.3|^6"
+		"symfony/yaml": "~4.4|~5.4|^6",
+		"ext-zip": "*"
 	},
 	"suggest": {
 		"doctrine/data-fixtures": "*"
@@ -55,7 +57,7 @@
 			"dev-master": "1.9.x-dev"
 		},
 		"symfony": {
-			"require": "~4.4|~5.3|^6"
+			"require": "~4.4|~5.4|^6"
 		}
 	},
 	"config": {
