@@ -14,15 +14,16 @@ use Doctrine\Persistence\ObjectRepository;
  */
 abstract class BaseGeonamesPostalCodeData implements FixtureInterface {
 
-	protected $batchSize = 1000;
+	protected int $batchSize = 1000;
 
 	// TODO remove as soon as doctrine/persistence >= 2.0 is required
-	protected final function _getRepository($manager) {
+	final protected function _getRepository($manager) {
 		return $manager->getRepository(GeoPostalCode::class);
 	}
 
 	// TODO remove as soon as doctrine/persistence >= 2.0 is required
-	protected final function _clearPostalCodesTable($manager) {
+	final protected function _clearPostalCodesTable($manager): void
+    {
 		foreach ($this->_getRepository($manager)->findAll() as $entity) {
 			$manager->remove($entity);
 		}
@@ -30,54 +31,55 @@ abstract class BaseGeonamesPostalCodeData implements FixtureInterface {
 	}
 
 	// TODO remove as soon as doctrine/persistence >= 2.0 is required
-	protected final function _addEntries($manager, $filename) {
+	final protected function _addEntries($manager, $filename): int
+    {
 		$repo = $this->_getRepository($manager);
 
 		$entriesAdded = 0;
 		$currentBatchEntries = [];
 
 		$fcontents = file($filename);
-		for ($i = 0, $numLines = count($fcontents); $i < $numLines; ++$i) {
-			$line = trim($fcontents[$i]);
-			$arr = explode("\t", $line);
+        foreach ($fcontents as $i => $iValue) {
+            $line = trim($iValue);
+            $arr = explode("\t", $line);
 
-			// skip if no lat/lng values
-			if (!array_key_exists(9, $arr) || !array_key_exists(10, $arr)) {
-				continue;
-			}
+            // skip if no lat/lng values
+            if (!array_key_exists(9, $arr) || !array_key_exists(10, $arr)) {
+                continue;
+            }
 
-			$country = $arr[0];
-			$postalCode = $arr[1];
+            $country = $arr[0];
+            $postalCode = $arr[1];
 
-			// skip duplicate entries in current batch
-			if (in_array($country.'-'.$postalCode, $currentBatchEntries, true)) {
-				continue;
-			}
+            // skip duplicate entries in current batch
+            if (in_array($country.'-'.$postalCode, $currentBatchEntries, true)) {
+                continue;
+            }
 
-			// skip duplicate entries already persisted
-			if ($repo->findOneBy(['country' => $country, 'postalCode' => $postalCode]) !== null) {
-				continue;
-			}
+            // skip duplicate entries already persisted
+            if ($repo->findOneBy(['country' => $country, 'postalCode' => $postalCode]) !== null) {
+                continue;
+            }
 
-			$entity = new GeoPostalCode();
-			$entity->setCountry($country);
-			$entity->setPostalCode($postalCode);
-			$entity->setLat((float) $arr[9]);
-			$entity->setLng((float) $arr[10]);
-			$manager->persist($entity);
+            $entity = new GeoPostalCode();
+            $entity->setCountry($country);
+            $entity->setPostalCode($postalCode);
+            $entity->setLat((float) $arr[9]);
+            $entity->setLng((float) $arr[10]);
+            $manager->persist($entity);
 
-			++$entriesAdded;
-			$currentBatchEntries[] = $country.'-'.$postalCode;
+            ++$entriesAdded;
+            $currentBatchEntries[] = $country.'-'.$postalCode;
 
-			if ((($i + 1) % $this->batchSize) === 0) {
-				$manager->flush();
-				$manager->clear();
-				$currentBatchEntries = [];
-				echo '.'; // progress indicator
-			}
-		}
+            if ((($i + 1) % $this->batchSize) === 0) {
+                $manager->flush();
+                $manager->clear();
+                $currentBatchEntries = [];
+                echo '.'; // progress indicator
+            }
+        }
 
-		$manager->flush(); // Flush for the last batch, which doesn't reach the batch size in most cases. (fixes #2)
+        $manager->flush(); // Flush for the last batch, which doesn't reach the batch size in most cases. (fixes #2)
 
 		echo ' ', $entriesAdded, "\n";
 
@@ -98,23 +100,24 @@ if (interface_exists(ObjectManager::class)) {
 		 * @param ObjectManager $manager
 		 * @return ObjectRepository
 		 */
-		protected function getRepository(ObjectManager $manager) {
+		protected function getRepository(ObjectManager $manager): ObjectRepository
+        {
 			return $this->_getRepository($manager);
 		}
 
 		/**
 		 * @param ObjectManager $manager
 		 */
-		protected function clearPostalCodesTable(ObjectManager $manager) {
+		protected function clearPostalCodesTable(ObjectManager $manager): void
+        {
 			$this->_clearPostalCodesTable($manager);
 		}
 
 		/**
-		 * @param ObjectManager $manager
-		 * @param string $filename
 		 * @return int Number of entries actually added.
 		 */
-		protected function addEntries(ObjectManager $manager, $filename) {
+		protected function addEntries(ObjectManager $manager, string $filename): int
+        {
 			return $this->_addEntries($manager, $filename);
 		}
 	}
@@ -129,23 +132,24 @@ if (interface_exists(ObjectManager::class)) {
 		 * @param LegacyObjectManager $manager
 		 * @return LegacyObjectRepository
 		 */
-		protected function getRepository(LegacyObjectManager $manager) {
+		protected function getRepository(LegacyObjectManager $manager): LegacyObjectRepository
+        {
 			return $this->_getRepository($manager);
 		}
 
 		/**
 		 * @param LegacyObjectManager $manager
 		 */
-		protected function clearPostalCodesTable(LegacyObjectManager $manager) {
+		protected function clearPostalCodesTable(LegacyObjectManager $manager): void
+        {
 			$this->_clearPostalCodesTable($manager);
 		}
 
 		/**
-		 * @param LegacyObjectManager $manager
-		 * @param string $filename
 		 * @return int Number of entries actually added.
 		 */
-		protected function addEntries(LegacyObjectManager $manager, $filename) {
+		protected function addEntries(LegacyObjectManager $manager, string $filename): int
+        {
 			return $this->_addEntries($manager, $filename);
 		}
 	}
