@@ -3,8 +3,10 @@
 namespace Craue\GeoBundle\Doctrine\Query\Mysql;
 
 use Doctrine\ORM\Query\AST\Functions\FunctionNode;
+use Doctrine\ORM\Query\AST\Node;
 use Doctrine\ORM\Query\Lexer;
 use Doctrine\ORM\Query\Parser;
+use Doctrine\ORM\Query\QueryException;
 use Doctrine\ORM\Query\SqlWalker;
 
 /**
@@ -17,12 +19,15 @@ use Doctrine\ORM\Query\SqlWalker;
  */
 class GeoDistanceByPostalCode extends FunctionNode {
 
-	protected $countryOrigin;
-	protected $postalCodeOrigin;
-	protected $countryDestination;
-	protected $postalCodeDestination;
+	protected Node $countryOrigin;
+	protected Node $postalCodeOrigin;
+	protected Node $countryDestination;
+	protected Node $postalCodeDestination;
 
-	public function parse(Parser $parser) : void {
+    /**
+     * @throws QueryException
+     */
+    public function parse(Parser $parser) : void {
 		$parser->match(Lexer::T_IDENTIFIER);
 		$parser->match(Lexer::T_OPEN_PARENTHESIS);
 		$this->countryOrigin = $parser->StringPrimary();
