@@ -11,68 +11,62 @@ use Symfony\Component\Validator\Constraints as Assert;
  */
 class GeoPostalCode {
 
-	/**
-	 * @var int
-	 */
-	protected $id;
+	protected int $id;
 
-	/**
-	 * @var string
-	 * @Assert\NotBlank
-	 */
-	protected $country;
+    #[Assert\NotBlank]
+	protected string $country;
 
-	/**
-	 * @var string
-	 * @Assert\NotBlank
-	 */
-	protected $postalCode;
+    #[Assert\NotBlank]
+	protected string $postalCode;
 
-	/**
-	 * @var double
-	 * @Assert\NotBlank
-	 */
-	protected $lat;
+    #[Assert\NotBlank]
+	protected float $lat;
 
-	/**
-	 * @var double
-	 * @Assert\NotBlank
-	 */
-	protected $lng;
+    #[Assert\NotBlank]
+	protected float $lng;
 
-	public function getId() {
+	public function getId(): int
+    {
 		return $this->id;
 	}
 
-	public function setCountry($country) {
+	public function setCountry(string $country): void
+    {
 		$this->country = $country;
 	}
 
-	public function getCountry() {
+	public function getCountry(): string
+    {
 		return $this->country;
 	}
 
-	public function setPostalCode($postalCode) {
+	public function setPostalCode(string $postalCode): void
+    {
 		$this->postalCode = $postalCode;
 	}
 
-	public function getPostalCode() {
+	public function getPostalCode(): string
+    {
 		return $this->postalCode;
 	}
 
-	public function setLat($lat) {
+	public function setLat(float $lat): void
+    {
 		$this->lat = $lat;
 	}
 
-	public function getLat() {
+	public function getLat(): float
+    {
 		return $this->lat;
 	}
 
-	public function setLng($lng) {
+	public function setLng(float $lng): void
+    {
 		$this->lng = $lng;
 	}
 
-	public function getLng() {
+	public function getLng(): float
+    {
 		return $this->lng;
 	}
 
