@@ -16,7 +16,8 @@ class GeoDistanceByPostalCodeTest extends IntegrationTestCase {
 	/**
 	 * @dataProvider getPlatformConfigs
 	 */
-	public function testNoResults($platform, $config, $requiredExtension) {
+	public function testNoResults($platform, $config, $requiredExtension): void
+    {
 		$this->initClient($requiredExtension, ['environment' => $platform, 'config' => $config]);
 
 		$result = $this->getPoisPerGeoDistanceByPostalCode('DE', '10551');
@@ -27,7 +28,8 @@ class GeoDistanceByPostalCodeTest extends IntegrationTestCase {
 	/**
 	 * @dataProvider getPlatformConfigs
 	 */
-	public function testDistance($platform, $config, $requiredExtension) {
+	public function testDistance($platform, $config, $requiredExtension): void
+    {
 		$this->initClient($requiredExtension, ['environment' => $platform, 'config' => $config]);
 
 		$this->persistGeoPostalCode('DE', '14473', 52.392759, 13.065135);
@@ -43,7 +45,8 @@ class GeoDistanceByPostalCodeTest extends IntegrationTestCase {
 	/**
 	 * @dataProvider getPlatformConfigs
 	 */
-	public function testUnknownPostalCode_withRadius($platform, $config, $requiredExtension) {
+	public function testUnknownPostalCode_withRadius($platform, $config, $requiredExtension): void
+    {
 		$this->initClient($requiredExtension, ['environment' => $platform, 'config' => $config]);
 
 		$this->persistGeoPostalCode('DE', '14473', 52.392759, 13.065135);
@@ -57,7 +60,8 @@ class GeoDistanceByPostalCodeTest extends IntegrationTestCase {
 	/**
 	 * @dataProvider getPlatformConfigs
 	 */
-	public function testUnknownPostalCode_withoutRadius($platform, $config, $requiredExtension) {
+	public function testUnknownPostalCode_withoutRadius($platform, $config, $requiredExtension): void
+    {
 		$this->initClient($requiredExtension, ['environment' => $platform, 'config' => $config]);
 
 		$this->persistGeoPostalCode('DE', '14473', 52.392759, 13.065135);
